refactor: split authority recovery responsibilities
