+++ packages/cli/test/production-authority-canonical-ingress/canonical-ingress.test.ts
@@ -1,36 +1,29 @@
 // harness-test-tier: integration
 import assert from "node:assert/strict";
 import { execFileSync } from "node:child_process";
-import { sign } from "node:crypto";
-import { existsSync, mkdirSync, mkdtempSync, readFileSync, realpathSync, rmSync, writeFileSync } from "node:fs";
-import { hostname, tmpdir } from "node:os";
+import { existsSync, readFileSync, rmSync } from "node:fs";
 import path from "node:path";
 import test from "node:test";
 import {
   channelDigest32,
-  connectionGeneration,
-  openLocalAuthorityKeyStore
+  connectionGeneration
 } from "../../../daemon/src/index.ts";
-import { createAuthorityKeyRegistryV1, firstPinAuthorityKeyV1 } from "../../../application/src/index.ts";
 import {
   decisionEntityId,
   createTaskPackagePath,
-  executionDeclaration,
   moduleEntityId,
   sha256Text,
   taskEntityId
 } from "../../../kernel/src/index.ts";
 import { defaultCliAdapterProvider } from "../../src/composition/adapter-registry.ts";
-import type { EntityId, ExecutionRecord } from "../../../kernel/src/index.ts";
+import type { EntityId } from "../../../kernel/src/index.ts";
 import type { ParsedCommand } from "../../src/cli/types.ts";
 import { daemonActorAttribution } from "../../src/composition/actor-attribution.ts";
 import { parseRecordArgs } from "../../src/cli/parsers/record.ts";
 import { parseNewTaskArgs } from "../../src/cli/parsers/new-task.ts";
 import { createCliCommandService } from "../../src/daemon/command-service.ts";
-import { authorityNamespaceProofBytes } from "../../src/daemon/authority-production-state.ts";
 import { createGitCanonicalPublicationInspector } from "../../src/daemon/authority-publication-evidence.ts";
 import { createProductionAuthorityLifecycle } from "../../src/daemon/production-authority-lifecycle.ts";
-import { openDurableAuthorityServiceState } from "../../src/daemon/authority-service-state.ts";
 import {
   defaultDaemonUserRoot,
   pollUntil,
@@ -39,6 +32,13 @@ import {
   runRawJsonMaybeFail,
   stopDaemon
 } from "../helpers/daemon-cli.ts";
+import {
+  authorityEventBodies,
+  authorityOperationRecords,
+  createFixture,
+  git,
+  latestAuthorityOperation
+} from "./fixture.ts";
 
 test("production service route preserves progress dry-run and publishes canonical task writes", { timeout: 60_000 }, async () => {
   const fixture = createFixture();
@@ -166,96 +166,6 @@ test("production service route preserves progress dry-run and publishes canonica
   } finally {
     await stopDaemon(fixture.repoRoot, userRoot).catch(() => undefined);
     if (process.env.KEEP_AUTHORITY_SERVICE_FIXTURE !== "1") rmSync(fixture.root, { recursive: true, force: true });
-  }
-});
-
-test("production service exposes its socket while authority recovery scans history, then uses the persisted increment", { timeout: 120_000 }, async () => {
-  const fixture = createFixture();
-  const userRoot = defaultDaemonUserRoot(fixture.root);
-  const env = {
-    HARNESS_ACTOR: "agent:codex",
-    HARNESS_DAEMON_MODE: "local",
-    HARNESS_DAEMON_USER_ROOT: userRoot,
-    HARNESS_DAEMON_IDLE_MS: "60000",
-    HARNESS_DAEMON_AUTOSTART_TIMEOUT_MS: "20000",
-    CODEX_THREAD_ID: "service-recovery-session"
-  };
-  const watermarkPath = path.join(
-    fixture.serviceRoot,
-    "authority",
-    Buffer.from("canonical", "utf8").toString("base64url"),
-    "recovery-watermark.json"
-  );
-  try {
-    for (let index = 0; index < 800; index += 1) {
-      git(fixture.authoredRoot, "commit", "-q", "--allow-empty", "-m", `fixture history ${index}`);
-    }
-    const seededState = openDurableAuthorityServiceState({ serviceStateRoot: fixture.serviceRoot, repoId: "canonical" });
-    await seededState.operationRegistry.put(indeterminateWithoutPublication());
-    await seededState.close();
-    const registered = runDaemonCommand(fixture.repoRoot, [
-      "daemon", "repo", "register", "--repo-id", "canonical", "--canonical-root", fixture.repoRoot,
-      "--user-root", userRoot, "--no-link", "--json"
-    ], env);
-    assert.equal(registered.ok, true, JSON.stringify(registered));
-
-    const coldStartedAt = Date.now();
-    runDaemonCommand(fixture.repoRoot, [
-      "daemon", "start", "--service", "--authority-manifest", fixture.manifestPath, "--json"
-    ], env);
-    const coldSocketMs = Date.now() - coldStartedAt;
-    const statusDuringRecovery = runDaemonCommand(
-      fixture.repoRoot,
-      ["daemon", "status", "--user-root", userRoot, "--json"],
-      env
-    );
-    assert.equal(statusDuringRecovery.reachable, true, JSON.stringify(statusDuringRecovery));
-    assert.equal(existsSync(watermarkPath), false, "cold full scan must still be in progress when the socket is first reachable");
-
-    const gated = runRawJsonMaybeFail(fixture.repoRoot, [
-      "task", "progress", "append", "task_01KXQ4WTA7Q4XJ5GDDRS1YXNG4", "--text", "must wait for recovery"
-    ], env);
-    assert.notEqual(gated.status, 0, JSON.stringify(gated.receipt));
-    assert.match(JSON.stringify(gated.receipt), /AUTHORITY_RECOVERY_IN_PROGRESS/u);
-
-    await pollUntil(
-      () => existsSync(watermarkPath) ? JSON.parse(readFileSync(watermarkPath, "utf8")) as { readonly commitSha?: string } : undefined,
-      (watermark) => watermark?.commitSha === git(fixture.authoredRoot, "rev-parse", "HEAD"),
-      (watermark, error) => JSON.stringify({ watermark, error: String(error ?? "") }),
-      { timeoutMs: 30_000 }
-    );
-    const coldRecoveryMs = Date.now() - coldStartedAt;
-    assert.equal(authorityOperationRecords(fixture.serviceRoot).find((record) => record.opId === "namespace-production:unpublished-startup")?.state, "REJECTED");
-    const committed = runRawJsonMaybeFail(fixture.repoRoot, [
-      "task", "progress", "append", "task_01KXQ4WTA7Q4XJ5GDDRS1YXNG4", "--text", "recovery completed"
-    ], env);
-    assert.equal(committed.status, 0, JSON.stringify(committed.receipt));
-    assert.equal(committed.receipt.ok, true, JSON.stringify(committed.receipt));
-
-    await stopDaemon(fixture.repoRoot, userRoot);
-    for (let index = 0; index < 5; index += 1) {
-      git(fixture.authoredRoot, "commit", "-q", "--allow-empty", "-m", `fixture increment ${index}`);
-    }
-    const incrementalHead = git(fixture.authoredRoot, "rev-parse", "HEAD");
-    const incrementalStartedAt = Date.now();
-    runDaemonCommand(fixture.repoRoot, [
-      "daemon", "start", "--service", "--authority-manifest", fixture.manifestPath, "--json"
-    ], env);
-    const incrementalSocketMs = Date.now() - incrementalStartedAt;
-    await pollUntil(
-      () => JSON.parse(readFileSync(watermarkPath, "utf8")) as { readonly commitSha?: string },
-      (watermark) => watermark.commitSha === incrementalHead,
-      (watermark, error) => JSON.stringify({ watermark, error: String(error ?? "") }),
-      { timeoutMs: 10_000 }
-    );
-    const incrementalRecoveryMs = Date.now() - incrementalStartedAt;
-
-    assert.ok(coldSocketMs < coldRecoveryMs, JSON.stringify({ coldSocketMs, coldRecoveryMs }));
-    assert.ok(incrementalRecoveryMs < coldRecoveryMs, JSON.stringify({ incrementalRecoveryMs, coldRecoveryMs }));
-    console.log(JSON.stringify({ coldSocketMs, coldRecoveryMs, incrementalSocketMs, incrementalRecoveryMs }));
-  } finally {
-    await stopDaemon(fixture.repoRoot, userRoot).catch(() => undefined);
-    rmSync(fixture.root, { recursive: true, force: true });
   }
 });
 
@@ -547,239 +457,3 @@ test("production canonical ingress accepts and journals one write for every cano
     rmSync(fixture.root, { recursive: true, force: true });
   }
 });
-
-function createFixture() {
-  const root = realpathSync(mkdtempSync(path.join(tmpdir(), "ha-production-canonical-ingress-")));
-  const repoRoot = path.join(root, "repo");
-  const authoredRoot = path.join(repoRoot, "harness");
-  const auxiliaryRoot = path.join(root, "auxiliary");
-  const auxiliaryAuthoredRoot = path.join(auxiliaryRoot, "harness");
-  const serviceRoot = path.join(root, "service-state");
-  const keyStateDirectory = path.join(serviceRoot, "keys/canonical");
-  mkdirSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0"), { recursive: true });
-  mkdirSync(serviceRoot, { recursive: true, mode: 0o700 });
-  writeFileSync(path.join(authoredRoot, "harness.yaml"), "schema: harness-anything/v1\nproject: production-ingress\n");
-  writeFileSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0/INDEX.md"), taskIndexBody("task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0"));
-  writeFileSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0/closeout.md"), "# Closeout\n\nProduction fixture qualified.\n");
-  mkdirSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG4"), { recursive: true });
-  writeFileSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG4/INDEX.md"), taskIndexBody("task_01KXQ4WTA7Q4XJ5GDDRS1YXNG4"));
-  mkdirSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG8-production-route"), { recursive: true });
-  writeFileSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG8-production-route/INDEX.md"), taskIndexBody("task_01KXQ4WTA7Q4XJ5GDDRS1YXNG8"));
-  const actor = {
-    personId: "person_alice",
-    displayName: "Alice",
-    primaryEmail: "alice@example.test",
-    providerId: "transport-derived/v1",
-    resolvedCredential: {
-      kind: "unix-socket-owner-boundary" as const,
-      issuer: `host:${hostname()}`,
-      subject: String(process.getuid?.() ?? 0)
-    }
-  };
-  const submittedExecution: ExecutionRecord = {
-    schema: "execution/v2", execution_id: "exe_01KXQ4WTA7Q4XJ5GDDRS1YXNG5", task_ref: "task/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0", state: "submitted",
-    primary_actor: { principal: { personId: "person_alice" }, executor: { kind: "agent", id: "codex" }, responsibleHuman: "person_alice" },
-    claimed_at: "2026-07-17T00:00:00.000Z", submitted_at: "2026-07-17T00:01:00.000Z", closed_at: null,
-    session_bindings: [], outputs: [{ evidence_id: "evidence:ingress", execution_ref: "execution/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0/exe_01KXQ4WTA7Q4XJ5GDDRS1YXNG5", locator: { substrate: "inline", text: "passed" } }],
-    submission: { completion_claim: "Ingress qualified", deliverables: ["journal"], evidence_refs: ["evidence:ingress"], verification_notes: ["integration"], known_gaps: [], residual_risks: [] }
-  };
-  mkdirSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0/executions"), { recursive: true });
-  writeFileSync(path.join(authoredRoot, "tasks/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG0/executions/exe_01KXQ4WTA7Q4XJ5GDDRS1YXNG5.md"), executionDeclaration.documentCodec.encode(submittedExecution));
-  const transcriptPath = path.join(root, "session-transcript.md");
-  writeFileSync(transcriptPath, `${JSON.stringify({
-    timestamp: "2026-07-17T00:00:00.000Z",
-    type: "event_msg",
-    payload: { type: "user_message", message: "Production session ingress." }
-  })}\n`);
-  writeFileSync(path.join(authoredRoot, "people.yaml"), [
-    "schema: harness-people/v1", "people:", "  - personId: person_alice", "    displayName: Alice",
-    "    primaryEmail: alice@example.test", "    roles: [owner]", "    credentials:",
-    "      - kind: unix-socket-owner-boundary", `        issuer: host:${hostname()}`,
-    `        subject: ${process.getuid?.() ?? 0}`, "roles:", "  - roleId: owner",
-    "    commandClasses: [admin, repo-write, repo-read, arbiter]", ""
-  ].join("\n"));
-  const keyStore = openLocalAuthorityKeyStore({
-    serviceStateRoot: serviceRoot,
-    stateDirectory: keyStateDirectory,
-    workspaceRoot: repoRoot,
-    authorityId: "authority.production",
-    issuer: "authority.production"
-  });
-  const now = Date.now();
-  const prepublished = keyStore.createPrepublishedKey({ generation: 1, nowMs: now - 1_000 });
-  const prepublishedRegistry = createAuthorityKeyRegistryV1({
-    authorityId: "authority.production", generation: 1, globalRevocationEpoch: 1, revision: 1,
-    entries: [prepublished]
-  });
-  const registry = firstPinAuthorityKeyV1({
-    registry: prepublishedRegistry,
-    keyId: prepublished.keyId,
-    expectedPinnedKeyId: prepublished.keyId,
-    pinEvidence: "fixture-out-of-band-pin",
-    verifierAcknowledgement: "fixture-verifier-ack",
-    activatedAtMs: now - 999
-  });
-  const registryPath = path.join(authoredRoot, "authority-key-registry.json");
-  writeFileSync(registryPath, `${JSON.stringify(registry, null, 2)}\n`);
-  const unsignedNamespace = {
-    schema: "operation-namespace/v1" as const,
-    workspaceId: "workspace-production",
-    deviceId: "device-production",
-    authorityGeneration: 1n,
-    namespaceId: "namespace-production",
-    expiresAt: BigInt(now + 60 * 60_000),
-    issuer: "authority.production",
-    keyId: prepublished.keyId
-  };
-  const proof = sign(null, authorityNamespaceProofBytes(unsignedNamespace), keyStore.signingProfile(registry, now).privateKey);
-  const manifestPath = path.join(serviceRoot, "authority-production.json");
-  writeFileSync(manifestPath, `${JSON.stringify({
-    schema: "authority-production-composition/v1",
-    serviceStateRoot: serviceRoot,
-    repos: [{
-      repoId: "canonical", canonicalRoot: repoRoot, workspaceId: "workspace-production",
-      deviceId: "device-production", viewId: "view-production", sessionId: "session-production",
-      authorityId: "authority.production", issuer: "authority.production", keyRegistryPath: registryPath,
-      keyStateDirectory, schemaTuple: productionTuple(), authorityGeneration: 1,
-      revocationEpochs: { global: "1", workspace: "1", device: "1", view: "1", principal: "1", executor: "1" },
-      admissionTokenRef: "admission-production", allowedExecutorAgentIds: ["codex"],
-      operationNamespace: {
-        ...unsignedNamespace,
-        authorityGeneration: unsignedNamespace.authorityGeneration.toString(),
-        expiresAt: unsignedNamespace.expiresAt.toString(),
-        proof: proof.toString("base64url")
-      }
-    }]
-  }, null, 2)}\n`);
-  writeFileSync(path.join(repoRoot, "README.md"), "# Distinct public repository\n");
-  git(repoRoot, "init", "-q");
-  git(repoRoot, "add", "README.md");
-  git(repoRoot, "commit", "-q", "-m", "seed distinct public fixture");
-  const publicHead = git(repoRoot, "rev-parse", "HEAD");
-  git(authoredRoot, "init", "-q");
-  git(authoredRoot, "add", ".");
-  git(authoredRoot, "commit", "-q", "-m", "seed canonical ingress fixture");
-  mkdirSync(auxiliaryAuthoredRoot, { recursive: true });
-  writeFileSync(path.join(auxiliaryAuthoredRoot, "harness.yaml"), "schema: harness-anything/v1\nproject: auxiliary-ingress\n");
-  git(auxiliaryAuthoredRoot, "init", "-q");
-  git(auxiliaryAuthoredRoot, "add", ".");
-  git(auxiliaryAuthoredRoot, "commit", "-q", "-m", "seed auxiliary ingress fixture");
-  return { root, repoRoot, authoredRoot, auxiliaryRoot, serviceRoot, manifestPath, actor, transcriptPath, publicHead };
-}
-
-function latestAuthorityOperation(serviceRoot: string): {
-  readonly state?: string;
-  readonly opId?: string;
-  readonly commitSha?: string;
-  readonly receipt?: { readonly tag?: string };
-} {
-  const operationPath = path.join(
-    serviceRoot,
-    "authority",
-    Buffer.from("canonical", "utf8").toString("base64url"),
-    "operations.jsonl"
-  );
-  const rows = readFileSync(operationPath, "utf8").trim().split("\n")
-    .map((line) => JSON.parse(line) as { readonly table?: string; readonly value?: Record<string, unknown> })
-    .filter((row) => row.table === "operation" && row.value);
-  assert.ok(rows.length > 0, "service route must persist an authority operation");
-  return rows.at(-1)!.value as {
-    readonly state?: string;
-    readonly opId?: string;
-    readonly commitSha?: string;
-    readonly receipt?: { readonly tag?: string };
-  };
-}
-
-function authorityOperationRecords(serviceRoot: string): ReadonlyArray<{
-  readonly state?: string;
-  readonly opId?: string;
-}> {
-  const operationPath = path.join(
-    serviceRoot,
-    "authority",
-    Buffer.from("canonical", "utf8").toString("base64url"),
-    "operations.jsonl"
-  );
-  const latest = new Map<string, { readonly state?: string; readonly opId?: string }>();
-  for (const line of readFileSync(operationPath, "utf8").trim().split("\n").filter(Boolean)) {
-    const row = JSON.parse(line) as { readonly table?: string; readonly key?: string; readonly value?: { readonly state?: string; readonly opId?: string } };
-    if (row.table === "operation" && row.key && row.value) latest.set(row.key, row.value);
-  }
-  return [...latest.values()];
-}
-
-function authorityEventBodies(authoredRoot: string): ReadonlyArray<string> {
-  const eventRoot = path.join(authoredRoot, "authority-attribution-events/v2");
-  if (!existsSync(eventRoot)) return [];
-  return execFileSync("find", [eventRoot, "-type", "f"], { encoding: "utf8" })
-    .trim().split("\n").filter(Boolean)
-    .map((eventPath) => readFileSync(eventPath, "utf8"));
-}
-
-function taskIndexBody(taskId: string): string {
-  return [
-    "---", "schema: task-package/v2", `task_id: ${taskId}`, "title: Production ingress",
-    "lifecycle:", "  bindingSchema: lifecycle-binding/v1", "  engine: local", "  status: active",
-    "  ref: ", "  titleSnapshot: Production ingress", "  url: ",
-    "  bindingCreatedAt: 2026-07-17T00:00:00.000Z", `  bindingFingerprint: sha256:${"b".repeat(64)}`,
-    "packageDisposition: active", "vertical: default", "preset: default",
-    "provenance:", "  - {runtime: \"human\", sessionId: \"fixture\", boundAt: \"2026-07-17T00:00:00.000Z\"}",
-    "---", "", "# Production ingress", ""
-  ].join("\n");
-}
-
-function indeterminateWithoutPublication() {
-  return {
-    workspaceId: "workspace-production",
-    opId: "namespace-production:unpublished-startup",
-    semanticDigest: "a".repeat(64),
-    state: "INDETERMINATE" as const,
-    receipt: {
-      tag: "INDETERMINATE" as const,
-      workspaceId: "workspace-production",
-      opId: "namespace-production:unpublished-startup",
-      semanticDigest: "a".repeat(64),
-      reason: "PUBLICATION_OUTCOME_UNKNOWN:startup performance fixture"
-    },
-    authorityIntegrity: {
-      schema: "authority-operation-integrity/v2" as const,
-      semanticRequestDigest: "a".repeat(64),
-      semanticMutationSetDigest: "c".repeat(64),
-      mutationRegistryVersion: 1,
-      actorAxesBindingDigest: "d".repeat(64),
-      canonicalMutationSet: {
-        registryVersion: 1,
-        mutations: [{
-          entity: { registryVersion: 1, entityKind: "task", canonicalRef: "task/task_01KXQ4WTA7Q4XJ5GDDRS1YXNG4" },
-          action: { registryVersion: 1, action: "append" }
-        }]
-      }
-    },
-    recordedProtocol: {
-      kind: "semantic-mutation-envelope/v2" as const,
-      schemaTuple: productionTuple()
-    },
-    canonicalRequestEnvelope: "startup-performance-envelope"
-  };
-}
-
-function productionTuple() {
-  return {
-    wire: 2, event: 2, receipt: 2, digest: 2, policy: 2,
-    commandRegistry: 1, entityRegistry: 1, mutationRegistry: 1, localState: 1, applyJournal: 1
-  } as const;
-}
-
-function git(rootDir: string, ...args: ReadonlyArray<string>): string {
-  return execFileSync("git", ["-C", rootDir, ...args], {
-    encoding: "utf8",
-    env: {
-      ...process.env,
-      GIT_AUTHOR_NAME: "ZeyuLi",
-      GIT_AUTHOR_EMAIL: "33339424+FairladyZ625@users.noreply.github.com",
-      GIT_COMMITTER_NAME: "ZeyuLi",
-      GIT_COMMITTER_EMAIL: "33339424+FairladyZ625@users.noreply.github.com"
-    }
-  }).trim();
-}
